+++ src/Interface/DataProviderServiceInterface.php
@@ -5,8 +5,8 @@ namespace App\Interface;
 use App\Model\Invoices\AccountData;
 use App\Model\Invoices\ClientData;
 use App\Model\Invoices\PagedResult;
-use App\Model\Invoices\ProjectData;
-use App\Model\Invoices\WorklogData;
+use App\Model\Invoices\ProjectDataCollection;
+use App\Model\Invoices\WorklogDataCollection;
 use App\Model\Planning\PlanningData;
 use App\Model\SprintReport\SprintReportData;
 use App\Model\SprintReport\SprintReportProjects;
@@ -14,12 +14,9 @@ use App\Model\SprintReport\SprintReportVersions;
 
 interface DataProviderServiceInterface
 {
-    public function getPlanningData(): PlanningData;
+    public function getPlanningDataSprints(): PlanningData;
 
-    /**
-     * @return array<ProjectData>
-     */
-    public function getAllProjectData(): array;
+    public function getPlanningDataWeeks(): PlanningData;
 
     /**
      * @return array<ClientData>
@@ -33,12 +30,13 @@ interface DataProviderServiceInterface
 
     public function getIssuesDataForProjectPaged(string $projectId, int $startAt = 0, $maxResults = 50): PagedResult;
 
-    /** @return array<WorklogData> */
-    public function getWorklogDataForProject(string $projectId): array;
-
     public function getSprintReportData(string $projectId, string $versionId): SprintReportData;
 
     public function getSprintReportProjects(): SprintReportProjects;
 
-    public function getSprintReportProjectVersions(string $projectId): SprintReportVersions;
+    public function getSprintReportVersions(string $projectId): SprintReportVersions;
+
+    public function getProjectDataCollection(): ProjectDataCollection;
+
+    public function getWorklogDataCollection(string $projectId): WorklogDataCollection;
 }