+++ src/Model/Planning/PlanningFormData.php
@@ -7,4 +7,5 @@ use App\Entity\DataProvider;
 class PlanningFormData
 {
     public DataProvider $dataProvider;
+    public string $viewType;
 }