+++ src/Service/LeantimeApiService.php
@@ -5,7 +5,13 @@ namespace App\Service;
 use App\Exception\ApiServiceException;
 use App\Exception\EconomicsException;
 use App\Interface\DataProviderServiceInterface;
+use App\Model\Invoices\IssueData;
 use App\Model\Invoices\PagedResult;
+use App\Model\Invoices\ProjectData;
+use App\Model\Invoices\ProjectDataCollection;
+use App\Model\Invoices\VersionData;
+use App\Model\Invoices\WorklogData;
+use App\Model\Invoices\WorklogDataCollection;
 use App\Model\Planning\Assignee;
 use App\Model\Planning\AssigneeProject;
 use App\Model\Planning\Issue;
@@ -13,6 +19,7 @@ use App\Model\Planning\PlanningData;
 use App\Model\Planning\Project;
 use App\Model\Planning\Sprint;
 use App\Model\Planning\SprintSum;
+use App\Model\Planning\Weeks;
 use App\Model\SprintReport\SprintReportData;
 use App\Model\SprintReport\SprintReportEpic;
 use App\Model\SprintReport\SprintReportIssue;
@@ -23,6 +30,7 @@ use App\Model\SprintReport\SprintReportVersion;
 use App\Model\SprintReport\SprintReportVersions;
 use App\Model\SprintReport\SprintStateEnum;
 use Doctrine\Common\Collections\ArrayCollection;
+use Symfony\Component\Uid\Ulid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 class LeantimeApiService implements DataProviderServiceInterface
@@ -42,29 +50,38 @@ class LeantimeApiService implements DataProviderServiceInterface
     ) {
     }
 
-    public function getAllProjectData(): array
+    public function getEndpoints(): array
     {
-        throw new \Exception('Not implemented');
+        return [
+            'base' => $this->leantimeUrl,
+        ];
     }
 
-    public function getClientDataForProject(string $projectId): array
+    /**
+     * Get all projects, including archived.
+     *
+     * @throws ApiServiceException|EconomicsException
+     */
+    public function getAllProjects(): mixed
     {
-        throw new \Exception('Not implemented');
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getAll', []);
     }
 
-    public function getAllAccountData(): array
+    /**
+     * Get all worklogs for project.
+     *
+     * @param $projectId
+     * @param string $from
+     * @param string $to
+     *
+     * @return mixed
+     *
+     * @throws ApiServiceException
+     * @throws EconomicsException
+     */
+    public function getProjectWorklogs($projectId, string $from = '2000-01-01', string $to = '3000-01-01'): mixed
     {
-        throw new \Exception('Not implemented');
-    }
-
-    public function getIssuesDataForProjectPaged(string $projectId, int $startAt = 0, $maxResults = 50): PagedResult
-    {
-        throw new \Exception('Not implemented');
-    }
-
-    public function getWorklogDataForProject(string $projectId): array
-    {
-        throw new \Exception('Not implemented');
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.timesheets.getAll', ['invEmpl' => '-1', 'invComp' => '-1', 'paid' => '-1', 'id' => $projectId]);
     }
 
     /**
@@ -72,7 +89,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      *
      * @return SprintReportProjects array of SprintReportProjects
      *
-     * @throws ApiServiceException
+     * @throws ApiServiceException|EconomicsException
      */
     public function getSprintReportProjects(): SprintReportProjects
     {
@@ -80,15 +97,109 @@ class LeantimeApiService implements DataProviderServiceInterface
         $projects = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getAll', []);
 
         foreach ($projects as $project) {
-            $sprintReportProjects->projects->add(
-                new SprintReportProject(
-                    $project->id,
-                    $project->name
-                )
-            );
+            $sprintReportProject = new SprintReportProject();
+            $sprintReportProject->id = $project->id;
+            $sprintReportProject->name = $project->name;
+
+            $sprintReportProjects->projects->add($sprintReportProject);
         }
 
         return $sprintReportProjects;
+    }
+
+    /**
+     * Get projectV2.
+     *
+     * @param string $projectId A project key or id
+     *
+     * @return SprintReportProject SprintReportProject
+     *
+     * @throws ApiServiceException|EconomicsException
+     */
+    public function getSprintReportProject(string $projectId): SprintReportProject
+    {
+        $project = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getProject', ['id' => $projectId]);
+
+        $sprintReportProject = new SprintReportProject();
+        $sprintReportProject->id = $project->id;
+        $sprintReportProject->name = $project->name;
+
+        return $sprintReportProject;
+    }
+
+    /**
+     * @throws ApiServiceException
+     * @throws EconomicsException
+     */
+    private function getProjectIssuesPaged($projectId, $startAt, $maxResults = 50): array
+    {
+        // TODO: Implement pagination.
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId]]);
+    }
+
+    /**
+     * @throws ApiServiceException
+     * @throws EconomicsException
+     */
+    public function getIssuesDataForProjectPaged(string $projectId, $startAt = 0, $maxResults = 50): PagedResult
+    {
+        $result = [];
+
+        $issues = $this->getProjectIssuesPaged($projectId, $startAt, $maxResults);
+
+        // Filter out all worklogs that do not belong to the project.
+        // TODO: Remove filter when issues are filtered correctly by projectId in the API.
+        $issues = array_filter($issues, fn ($issue) => $issue->projectId == $projectId);
+
+        foreach ($issues as $issue) {
+            $issueData = new IssueData();
+
+            $issueData->name = $issue->headline;
+            $issueData->status = $issue->status;
+            $issueData->projectTrackerId = $issue->id;
+            // Leantime does not have a key for each issue.
+            $issueData->projectTrackerKey = $issue->id;
+            $issueData->accountId = '';
+            $issueData->accountKey = '';
+            $issueData->epicKey = $issue->tags;
+            $issueData->epicName = $issue->tags;
+            if (isset($issue->milestoneid) && isset($issue->milestoneHeadline)) {
+                $issueData->versions?->add(new VersionData($issue->milestoneid, $issue->milestoneHeadline));
+            }
+            $issueData->projectId = $issue->projectId;
+            $result[] = $issueData;
+        }
+
+        return new PagedResult($result, $startAt, count($issues), count($issues));
+    }
+
+    public function getProjectDataCollection(): ProjectDataCollection
+    {
+        $projectDataCollection = new ProjectDataCollection();
+        $projects = $this->getAllProjects();
+
+        foreach ($projects as $project) {
+            $projectData = new ProjectData();
+            $projectData->name = $project->name;
+            $projectData->projectTrackerId = $project->id;
+            // Leantime does not have a key for each project.
+            $projectData->projectTrackerKey = $project->id;
+            $projectData->projectTrackerProjectUrl = $this->leantimeUrl.'#/tickets/showTicket/'.$project->id;
+
+            $projectVersions = $this->getSprintReportVersions($project->id);
+            foreach ($projectVersions as $projectVersion) {
+                $projectData->versions?->add($projectVersion);
+            }
+
+            $projectDataCollection->projectData->add($projectData);
+        }
+
+        return $projectDataCollection;
+    }
+
+    public function getClientDataForProject(string $projectId): array
+    {
+        throw new ApiServiceException('Method not implemented', 501);
     }
 
     /**
@@ -104,6 +215,11 @@ class LeantimeApiService implements DataProviderServiceInterface
         return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getProject', ['id' => $key]);
     }
 
+    public function getAllAccountData(): array
+    {
+        return [];
+    }
+
     /**
      * Get milestone.
      *
@@ -112,26 +228,61 @@ class LeantimeApiService implements DataProviderServiceInterface
      *
      * @throws ApiServiceException
      */
-    private function getMilestone($key): mixed
+    public function getMilestone($key): mixed
     {
         return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getTicket', ['id' => $key]);
     }
 
-    public function getSprintReportProjectVersions(string $projectId): SprintReportVersions
+    public function getSprintReportVersions(string $projectId): SprintReportVersions
     {
         $sprintReportVersions = new SprintReportVersions();
         $projectVersions = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAllMilestones', ['searchCriteria' => ['currentProject' => $projectId, 'type' => 'milestone']]);
 
         foreach ($projectVersions as $projectVersion) {
-            $sprintReportVersions->versions->add(
-                new SprintReportVersion(
-                    $projectVersion->id,
-                    $projectVersion->headline
-                )
-            );
+            $sprintReportVersion = new SprintReportVersion();
+            $sprintReportVersion->id = $projectVersion->id;
+            $sprintReportVersion->name = $projectVersion->headline;
+            $sprintReportVersion->projectTrackerId = $projectVersion->projectId;
+
+            $sprintReportVersions->versions->add($sprintReportVersion);
         }
 
         return $sprintReportVersions;
+    }
+
+    public function getWorklogDataCollection(string $projectId): WorklogDataCollection
+    {
+        $worklogDataCollection = new WorklogDataCollection();
+        $worklogs = $this->getProjectWorklogs($projectId);
+
+        $workersData = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.users.getAll');
+
+        $workers = array_reduce($workersData, function ($carry, $item) {
+            $carry[$item->id] = $item->username;
+
+            return $carry;
+        }, []);
+
+        // Filter out all worklogs that do not belong to the project.
+        // TODO: Remove filter when worklogs are filtered correctly by projectId in the API.
+        $worklogs = array_filter($worklogs, fn ($worklog) => $worklog->projectId == $projectId);
+
+        foreach ($worklogs as $worklog) {
+            $worklogData = new WorklogData();
+            if (isset($worklog->ticketId)) {
+                $worklogData->projectTrackerId = $worklog->id;
+                $worklogData->comment = $worklog->description ?? '';
+                $worklogData->worker = $workers[$worklog->userId] ?? $worklog->userId;
+                $worklogData->timeSpentSeconds = (int) ($worklog->hours * 60 * 60);
+                $worklogData->started = new \DateTime($worklog->workDate);
+                $worklogData->projectTrackerIsBilled = false;
+                $worklogData->projectTrackerIssueId = $worklog->ticketId;
+
+                $worklogDataCollection->worklogData->add($worklogData);
+            }
+        }
+
+        return $worklogDataCollection;
     }
 
     /**
@@ -155,13 +306,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      */
     public function getTicketsInSprint(string $sprintId): array
     {
-        $result = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', [
-            'searchCriteria' => [
-                'sprint' => $sprintId,
-            ],
-        ]);
-
-        return $result;
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['sprint' => $sprintId]]);
     }
 
     /**
@@ -169,7 +314,11 @@ class LeantimeApiService implements DataProviderServiceInterface
      */
     private function getIssueSprint($issueEntry): SprintReportSprint
     {
-        $sprint = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.sprints.getSprint', ['id' => $issueEntry->sprint]);
+        $sprint = false;
+
+        if ((bool) $issueEntry->sprint) {
+            $sprint = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.sprints.getSprint', ['id' => $issueEntry->sprint]);
+        }
 
         if ($sprint) {
             $sprintState = SprintStateEnum::OTHER;
@@ -201,12 +350,225 @@ class LeantimeApiService implements DataProviderServiceInterface
     }
 
     /**
+     * @throws ApiServiceException
+     */
+    private function getAllIssues(): array
+    {
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => []]);
+    }
+
+    /**
      * Create data for planning page.
      *
      * @throws ApiServiceException
      * @throws \Exception
      */
-    public function getPlanningData(): PlanningData
+    public function getPlanningDataWeeks(): PlanningData
+    {
+        $planning = new PlanningData();
+        $assignees = $planning->assignees;
+        $projects = $planning->projects;
+        $weeks = $planning->weeks;
+
+        $currentYear = (int) (new \DateTime())->format('Y');
+
+        for ($weekNumber = 1; $weekNumber <= 52; ++$weekNumber) {
+            $date = (new \DateTime())->setISODate($currentYear, $weekNumber);
+            $week = (int) $date->format('W'); // Cast as int to remove leading zero.
+            $weekFirstDay = $date->setISODate($currentYear, $week, 1)->format('j/n');
+            $weekLastDay = $date->setISODate($currentYear, $week, 5)->format('j/n');
+            $weekIsSupport = 1 === $week % 4;
+
+            if ($weekIsSupport) {
+                $supportWeek = new Weeks();
+                $supportWeek->weekCollection->add($week);
+                $supportWeek->weeks = 1;
+                $supportWeek->weekGoalLow = $this->weekGoalLow;
+                $supportWeek->weekGoalHigh = $this->weekGoalHigh;
+                $supportWeek->displayName = (string) $week;
+                $supportWeek->dateSpan = $weekFirstDay.' - '.$weekLastDay;
+                $weeks->add($supportWeek);
+            } else {
+                if (isset($regularWeek)) {
+                    $regularWeek->weekCollection->add($week);
+                    ++$regularWeek->weeks;
+                    $regularWeek->displayName .= '-'.$week;
+                    if (3 === count($regularWeek->weekCollection)) {
+                        $regularWeek->dateSpan .= ' - '.$weekLastDay;
+                        $weeks->add($regularWeek);
+                        unset($regularWeek);
+                    }
+                } else {
+                    $regularWeek = new Weeks();
+                    $regularWeek->weekCollection->add($week);
+                    $regularWeek->weeks = 1;
+                    $regularWeek->weekGoalLow = $this->weekGoalLow * 3;
+                    $regularWeek->weekGoalHigh = $this->weekGoalHigh * 3;
+                    $regularWeek->displayName = (string) $week;
+                    $regularWeek->dateSpan = $weekFirstDay;
+                }
+            }
+        }
+
+        $weekIssues = [];
+        $allIssues = $this->getAllIssues();
+
+        foreach ($allIssues as $issue) {
+            $issueYear = new \DateTime($issue->dateToFinish);
+            $issueYear = $issueYear->format('Y');
+
+            $issueWeek = new \DateTime($issue->dateToFinish);
+            $issueWeek = (int) $issueWeek->format('W');
+
+            if ('-0001' !== $issueYear) {
+                $weekIssues[$issueWeek][] = $issue;
+            } else {
+                $weekIssues['unscheduled'][] = $issue;
+            }
+        }
+        foreach ($weekIssues as $week => $issues) {
+            foreach ($issues as $issueData) {
+                if ('0' !== $issueData->status) { // excludes done issues.
+                    $week = (string) $week;
+                    $projectKey = (string) $issueData->projectId;
+                    $projectDisplayName = $issueData->projectName;
+
+                    $hoursRemaining = ($issueData->planHours - $issueData->bookedHours);
+                    if (empty($issueData->editorId)) {
+                        $assigneeKey = 'unassigned';
+                        $assigneeDisplayName = 'Unassigned';
+                    } else {
+                        $assigneeKey = (string) $issueData->editorId;
+                        if (isset($issueData->editorFirstname) || isset($issueData->editorLastname)) {
+                            $assigneeDisplayName = $issueData->editorFirstname.' '.$issueData->editorLastname;
+                        } else {
+                            $assigneeDisplayName = 'Name missing';
+                        }
+                    }
+                    // Add assignee if not already added.
+                    if (!$assignees->containsKey($assigneeKey)) {
+                        $assignees->set($assigneeKey, new Assignee($assigneeKey, $assigneeDisplayName));
+                    }
+
+                    /** @var Assignee $assignee */
+                    $assignee = $assignees->get($assigneeKey);
+
+                    // Add sprint if not already added.
+                    if (!$assignee->sprintSums->containsKey($week)) {
+                        $assignee->sprintSums->set($week, new SprintSum($week));
+                    }
+
+                    /** @var SprintSum $sprintSum */
+                    $sprintSum = $assignee->sprintSums->get($week);
+                    $sprintSum->sumHours += $hoursRemaining;
+
+                    // Add assignee project if not already added.
+                    if (!$assignee->projects->containsKey($projectKey)) {
+                        $assigneeProject = new AssigneeProject($projectKey, $projectDisplayName);
+                        $assignee->projects->set($projectKey, $assigneeProject);
+                    }
+
+                    /** @var AssigneeProject $assigneeProject */
+                    $assigneeProject = $assignee->projects->get($projectKey);
+
+                    // Add project sprint sum if not already added.
+                    if (!$assigneeProject->sprintSums->containsKey($week)) {
+                        $assigneeProject->sprintSums->set($week, new SprintSum($week));
+                    }
+
+                    /** @var SprintSum $projectSprintSum */
+                    $projectSprintSum = $assigneeProject->sprintSums->get($week);
+                    if (isset($projectSprintSum)) {
+                        $projectSprintSum->sumHours += $hoursRemaining;
+                    }
+
+                    $assigneeProject->issues->add(
+                        new Issue(
+                            $issueData->id,
+                            $issueData->headline,
+                            isset($issueData->hourRemaining) ? $hoursRemaining : null,
+                            $this->leantimeUrl.'/tickets/showKanban?showTicketModal='.$issueData->id.'#/tickets/showTicket/'.$issueData->id,
+                            $week
+                        )
+                    );
+
+                    // Add project if not already added.
+                    if (!$projects->containsKey($projectKey)) {
+                        $projects->set($projectKey, new Project(
+                            $projectKey,
+                            $projectDisplayName,
+                        ));
+                    }
+
+                    /** @var Project $project */
+                    $project = $projects->get($projectKey);
+
+                    // Add sprint sum if not already added.
+
+                    if (!$project->sprintSums->containsKey($week)) {
+                        $project->sprintSums->set($week, new SprintSum($week));
+                    }
+
+                    /** @var SprintSum $projectSprintSum */
+                    $projectSprintSum = $project->sprintSums->get($week);
+                    $projectSprintSum->sumHours += $hoursRemaining;
+
+                    if (!$project->assignees->containsKey($assigneeKey)) {
+                        $project->assignees->set($assigneeKey, new AssigneeProject(
+                            $assigneeKey,
+                            $assigneeDisplayName,
+                        ));
+                    }
+
+                    /** @var AssigneeProject $projectAssignee */
+                    $projectAssignee = $project->assignees->get($assigneeKey);
+
+                    if (!$projectAssignee->sprintSums->containsKey($week)) {
+                        $projectAssignee->sprintSums->set($week, new SprintSum($week));
+                    }
+
+                    /** @var SprintSum $projectAssigneeSprintSum */
+                    $projectAssigneeSprintSum = $projectAssignee->sprintSums->get($week);
+                    $projectAssigneeSprintSum->sumHours += $hoursRemaining;
+
+                    $projectAssignee->issues->add(new Issue(
+                        $issueData->id,
+                        $issueData->headline,
+                        isset($issueData->hourRemaining) ? $hoursRemaining : null,
+                        $this->leantimeUrl.'/tickets/showKanban?showTicketModal='.$issueData->id.'#/tickets/showTicket/'.$issueData->id,
+                        $week
+                    ));
+                }
+            }
+        }
+
+        // Sort assignees by name.
+        /** @var \ArrayIterator $iterator */
+        $iterator = $assignees->getIterator();
+        $iterator->uasort(function ($a, $b) {
+            return mb_strtolower($a->displayName) <=> mb_strtolower($b->displayName);
+        });
+        $planning->assignees = new ArrayCollection(iterator_to_array($iterator));
+
+        // Sort projects by name.
+        /** @var \ArrayIterator $iterator */
+        $iterator = $projects->getIterator();
+        $iterator->uasort(function ($a, $b) {
+            return mb_strtolower($a->displayName) <=> mb_strtolower($b->displayName);
+        });
+
+        $planning->projects = new ArrayCollection(iterator_to_array($iterator));
+
+        return $planning;
+    }
+
+    /**
+     * Create data for planning page.
+     *
+     * @throws ApiServiceException
+     * @throws \Exception
+     */
+    public function getPlanningDataSprints(): PlanningData
     {
         $planning = new PlanningData();
         $assignees = $planning->assignees;
@@ -379,9 +741,17 @@ class LeantimeApiService implements DataProviderServiceInterface
         return $planning;
     }
 
-    public function getTimesheetsForTicket($ticketId): mixed
+    public function getTimesheetsForTicket(string $ticketId): mixed
     {
         return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.timesheets.getAll', ['invEmpl' => '-1', 'invComp' => '-1', 'paid' => '-1', 'ticketFilter' => $ticketId]);
+    }
+
+    /**
+     * @throws ApiServiceException
+     */
+    private function getIssuesForProjectMilestone($projectId, $milestoneId): array
+    {
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId, 'milestone' => $milestoneId]]);
     }
 
     /**
@@ -428,18 +798,20 @@ class LeantimeApiService implements DataProviderServiceInterface
             $issue->epic = $tag;
 
             // Get sprint for issue.
-            try {
-                $issueSprint = $this->getIssueSprint($issueEntry);
+            if ((bool) $issueEntry->sprint) {
+                try {
+                    $issueSprint = $this->getIssueSprint($issueEntry);
 
-                if (!$sprints->containsKey($issueSprint->id)) {
-                    $sprints->set($issueSprint->id, $issueSprint);
+                    if (!$sprints->containsKey($issueSprint->id)) {
+                        $sprints->set($issueSprint->id, $issueSprint);
+                    }
+                    // Set which sprint the issue is assigned to.
+                    if (SprintStateEnum::ACTIVE === $issueSprint->state || SprintStateEnum::FUTURE === $issueSprint->state) {
+                        $issue->assignedToSprint = $issueSprint;
+                    }
+                } catch (ApiServiceException) {
+                    // Ignore if sprint is not found.
                 }
-                // Set which sprint the issue is assigned to.
-                if (SprintStateEnum::ACTIVE === $issueSprint->state || SprintStateEnum::FUTURE === $issueSprint->state) {
-                    $issue->assignedToSprint = $issueSprint;
-                }
-            } catch (ApiServiceException) {
-                // Ignore if sprint is not found.
             }
             $worklogs = $this->getTimesheetsForTicket($issueEntry->id);
 
@@ -522,14 +894,6 @@ class LeantimeApiService implements DataProviderServiceInterface
     }
 
     /**
-     * @throws ApiServiceException
-     */
-    private function getIssuesForProjectMilestone($projectId, $milestoneId): array
-    {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId, 'milestone' => $milestoneId]]);
-    }
-
-    /**
      * Get from Leantime.
      *
      * @throws ApiServiceException
@@ -542,7 +906,7 @@ class LeantimeApiService implements DataProviderServiceInterface
                 ['json' => [
                     'jsonrpc' => '2.0',
                     'method' => $method,
-                    'id' => '1',
+                    'id' => (new Ulid())->jsonSerialize(),
                     'params' => $params,
                 ]]
             );